Rethrow .catch errors
By .catching these promises (and others that are not yet fixed), this is
actually masking any error that is thrown, not just things that arise
from the promise in question failing. This causes a lot of issues. For
example, the test in app_action.spec.js that is changed in this commit
was not working as intended because the cfAPI.fetchApp call was never
being stubbed - and the .catch handler was masking the fact that the
cfAPI was actually trying to make a real call. Many other errors are
masked because of these .catch handlers which makes development much
harder.