Load email templates using govau/emailtemplate
The core reason behind this change is the hard-coded "Invitation to join
cloud.gov" subject string. Instead of asking users of this application
to provide configuration in the form of config files or environment
variables for the subject string, we take a different approach.

Instead, we use govau/emailtemplate which allows the developer to
provide 3 files: html.html, text.txt and subject.txt for every email
template they wish to send.

Each of these templates is a Go text/template (or html/template for
html.html). Thus each template has the power of Go's templating language
for rendering variables, using conditionals, loops, func maps, etc. This
makes it easy to internationalize templates, or switch templates for an
A/B test. Particularly worth mentioning is that the subject can now
contain variables, e.g. the user's name (if desired).